The C version of Pascal's Triangle
The C version of Pascal's Triangle, the first program homework in my program class. But it suppose to be a Python version and I failed to write. Then I decided to write a C then translate it to the Python. Unfortunaately, I still faced so much bugs when translating. Anyway I did it. I upload this to recognize all my effort to struggle with Python.
Fork u Python!